feat: status column moved member to chall(cover deleted)  and notified column added to chall
