--- server/server/src/main/java/com/cactusvilleage/server/auth/entities/Status.java
@@ -1,5 +1,0 @@
-package com.cactusvilleage.server.auth.entities;
-
-public enum Status {
-    SUCCESS, FAIL, IN_PROGRESS, NONE
-}
