+++ server/server/src/main/java/com/cactusvilleage/server/challenge/entities/Status.java
@@ -1,0 +1,5 @@
+package com.cactusvilleage.server.challenge.entities;
+
+public enum Status {
+    SUCCESS, FAIL, IN_PROGRESS, DELETED
+}
